pyproject.toml
[project]
name = "line-girl-effect"
version = "0.1.0"
description = "the 'Line Girl' effect"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pillow>=11.3.0",
    "scipy>=1.13.1",
]
